Refactor update check and update dependencies
Moved the update checking logic from utils.py to a new updater.py module and updated references in home_view.py. Updated dependencies in pyproject.toml, including flet and minecraft-launcher-lib versions. Updated copyright years and version strings across multiple files. Added startup screen configuration and improved get_app_path error handling.
