+++ src/widgets/app.py
@@ -1,3 +1,7 @@
+# This file is part of PYZ-LAUNCHER-FOR-MINECRAFT (https://github.com/ZadkielAvendano/PyZ-Launcher-for-Minecraft)
+# Copyright (c) 2026 Zadkiel Avendano and collaborators
+# License-Identifier: MIT License
+
 import flet as ft
 
 class WindowTittleBar(ft.AppBar):
